# perf: use precomputed zero hashes in process_subtree