Stacked on https://github.com/ralexstokes/ssz-rs/pull/169

A lot of leaves are empty and we don't need to hash them.
On the layer above we can do a similar thing; if both the leaves are zero hashes we can precompute their parent as well.

```
Prove Benchmark - File: benches/21315748.json - size 247/123
                        time:   [96.891 ms 97.480 ms 98.564 ms]
                        change: [-36.824% -36.084% -35.371%] (p = 0.00 < 0.05)
                        Performance has improved.

Benchmarking Prove Benchmark - File: benches/21327802.json - size 261/130: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.3s or enable flat sampling.
Prove Benchmark - File: benches/21327802.json - size 261/130
                        time:   [96.672 ms 97.006 ms 97.291 ms]
                        change: [-38.752% -37.874% -37.086%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
```