@thedevbirb 2024-12-09
Since PRs are stacked it is a bit difficult to see the new changes. Can you outline a changelog or specify what are the main new functions introduced and where are they used?